% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_participation.R
\docType{data}
\name{course_participation}
\alias{course_participation}
\title{course_participation: Participation in the Course}
\format{
A tibble with 734 rows and 5 variables
\describe{
  \item{duration_minutes}{Minutes spent in the session}
  \item{guest}{Whether the user is a guest or not. The course instructor and their notetaker is not a guest}
  \item{date}{Date of the session}
  \item{session}{Session number}
  \item{anonymized_name}{Anonymized hash of the name as an identifier}
}
}
\usage{
course_participation
}
\description{
Sessions attended and participated in, for the course (anonymized)
}
\keyword{datasets}
