% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds4owd_browser.R
\docType{data}
\name{ds4owd_browser}
\alias{ds4owd_browser}
\title{ds4owd_browser: Data on visits to the DS4OWD website by browser}
\format{
A tibble with 17 rows and 6 variables
\describe{
  \item{browser}{Browser used to accesss the DS4OWD website}
  \item{pageviews}{Number of pages viewed}
  \item{visitors}{Total number of unique visitors}
  \item{visit_duration}{Duration of visit (seconds)}
  \item{visits}{Number of non-unique visits to the website}
  \item{bounce_rate}{Percentage of users "bouncing" off website. A bounce is when a visitor spends less than 30 seconds on the website}
}
}
\usage{
ds4owd_browser
}
\description{
This dataset contains information on visits to the DS4OWD website by browser used.
}
\keyword{datasets}
