% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds4owd_country.R
\docType{data}
\name{ds4owd_country}
\alias{ds4owd_country}
\title{ds4owd_country: Data on visits to the DS4OWD website by country}
\format{
A tibble with 88 rows and 6 variables
\describe{
  \item{country}{Country of visit to the website}
  \item{pageviews}{Number of pages viewed}
  \item{visit_duration}{Duration of visit (seconds)}
  \item{visitors}{Total number of unique visitors}
  \item{visits}{Number of non-unique visits to the website}
  \item{bounce_rate}{Percentage of users "bouncing" off website. A bounce is when a visitor spends less than 30 seconds on the website}
}
}
\usage{
ds4owd_country
}
\description{
Visits to the DS4OWD website by country
}
\keyword{datasets}
