% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds4owd_source.R
\docType{data}
\name{ds4owd_source}
\alias{ds4owd_source}
\title{ds4owd_source: Data on source of website visits}
\format{
A tibble with 27 rows and 6 variables
\describe{
  \item{source}{Source of visit (Search engine, organic etc.)}
  \item{pageviews}{Number of pages viewed}
  \item{visit_duration}{Duration of visit (seconds)}
  \item{visitors}{Total number of unique visitors}
  \item{visits}{Number of non-unique visits to the website}
  \item{bounce_rate}{Percentage of users "bouncing" off website. A bounce is when a visitor spends less than 30 seconds on the website}
}
}
\usage{
ds4owd_source
}
\description{
Data on source of website visits
}
\keyword{datasets}
