% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds4owd_timeseries.R
\docType{data}
\name{ds4owd_timeseries}
\alias{ds4owd_timeseries}
\title{ds4owd_timeseries: Timeseries data on website traffic}
\format{
A tibble with 468 rows and 6 variables
\describe{
  \item{date}{Date of data collection}
  \item{pageviews}{Number of pages viewed}
  \item{visitors}{Total number of unique visitors}
  \item{visit_duration}{Duration of visit (seconds)}
  \item{visits}{Number of non-unique visits to the website}
  \item{bounce_rate}{Percentage of users "bouncing" off website. A bounce is when a visitor spends less than 30 seconds on the website}
}
}
\usage{
ds4owd_timeseries
}
\description{
Timeseries data on website traffic to the DS4OWD website
}
\keyword{datasets}
