% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\docType{data}
\name{locations}
\alias{locations}
\title{locations: Location of the course participants}
\format{
A tibble with 212 rows and 3 variables
\describe{
  \item{iso_country}{ISO Name for country of course participant}
  \item{country}{Name of country}
  \item{city}{Name of city}
}
}
\usage{
locations
}
\description{
Location by country (and city) of the course participants. This is anonymized
}
\keyword{datasets}
